+++ producers/eudrb/src/EUDRBProducer.cxx
@@ -219,7 +219,7 @@ public:
   }
   bool ReadoutEventOld(RawDataEvent & ev) {
     static bool readingstarted = false;
-    unsigned long total_bytes=0;
+    uint32_t total_bytes=0;
     std::cout << "--" << std::endl; 
     for (size_t i=0; i <= m_boards.size(); ++i) {
       size_t n_test = i < m_boards.size() ? i : m_master; // make sure master is read out last
@@ -241,7 +241,7 @@ public:
       }
 
       Timer t_mblt, t_reset;
-      unsigned long number_of_bytes=datasize*4;
+      uint32_t number_of_bytes=datasize*4;
 
       //printf("number of bytes = %ld\n",number_of_bytes);
       if (number_of_bytes <= 0) {
@@ -294,7 +294,7 @@ public:
   }
   bool ReadoutEventNew(RawDataEvent & ev) {
     static bool readingstarted = false;
-    unsigned long total_bytes=0;
+    uint32_t total_bytes=0;
     Timer t_wait;
     //std::cout << "ReadoutEventNew" << std::endl;
     if (!m_boards[m_boards.size()-1]->EventDataReady()) {
@@ -310,8 +310,8 @@ public:
       readingstarted = true;
     }
     //std::cout<< "--"<< std::endl;
-    unsigned long off = 0;
-    unsigned long pivot = 0;
+    uint32_t off = 0;
+    uint32_t pivot = 0;
     bool off_set = false;
     bool pivot_set = false;
 
@@ -355,7 +355,7 @@ public:
           off = m_buffer[3];
           off_set = true;
         } else {
-          unsigned long diff = (9216 + off - m_buffer[3]) % 9216;
+          uint32_t diff = (9216 + off - m_buffer[3]) % 9216;
           if(diff > 4 && diff < 9212) {
             EUDAQ_WARN("data consistency check 1 for board " + to_string(n_eudrb) +" in event " + to_string(m_ev) + " failed! The offset difference in pixel is " + to_string(diff) + "!");
 //            std::cout << "------ header -----" << std::endl; 
@@ -367,7 +367,7 @@ public:
           }
         }
       }
-      const unsigned long number_of_bytes = 4 * (m_buffer[0] & 0xFFFFF);
+      const uint32_t number_of_bytes = 4 * (m_buffer[0] & 0xFFFFF);
             
       if (doprint(m_ev)) std::cout << "DEBUG: read leading words board " << n_eudrb << ", remaining = " << number_of_bytes << std::endl;
 
@@ -385,7 +385,7 @@ public:
         if (doprint(m_ev)) std::cout << "OK" << std::endl;
         
 	if (m_boards[0]->Det() == "MIMOSA26") {
-          const unsigned long p = m_buffer[1] & 0x3FFF;
+          const uint32_t p = m_buffer[1] & 0x3FFF;
           if(!pivot_set)
             {
               pivot = p;
@@ -394,7 +394,7 @@ public:
           else
             { 
              
-            unsigned long diff = (9216 + pivot - p) % 9216;
+            uint32_t diff = (9216 + pivot - p) % 9216;
              
             if(diff > 4 && diff < 9212)
               {
